backend/requirements.txt
celery[redis] == 5.4.0
Django == 5.0.4
django-cachalot == 2.6.3
django-cors-headers == 4.3.1
django-elasticsearch-dsl == 8.0
django-elasticsearch-dsl-drf == 0.22.5
django-filter == 24.2
django-prometheus == 2.3.1
django-taggit == 5.0.1
djangorestframework == 3.15.1
djangorestframework-simplejwt == 5.3.1
drf-spectacular == 0.27.2
drf_redesign == 0.3.9
elasticsearch-dsl == 8.12.0
factory-boy == 3.3.1
faker == 28.0.0
flower == 2.0.1
gunicorn == 22.0.0
locust == 2.31.4
psycopg2 == 2.9.9
python-logstash == 0.4.8
rich == 13.7.1
ruff == 0.3.7
